pyproject.toml
# pyproject.toml
[project]
name = "cai-chess-agent"
version = "0.1.0"
description = "Chess.com profile analyzer with LangChain agent"
dependencies = [
    "chess==1.11.2",
    "langchain==0.3.26",
    "langchain-community==0.3.27",
    "langchain-openai==0.3.27",
    "openai==1.93.0",
    "python-dotenv==1.1.1",
    "requests==2.32.4",
    "stockfish==3.28.0",
    "tiktoken==0.9.0",
]
